task3/server/requirements.txt
Django~=5.0.2
psycopg2~=2.9.9
pillow~=10.2.0
celery~=5.3.6
redis~=5.0.1
requests~=2.31.0
djangorestframework~=3.14.0
drf-spectacular~=0.27.1
djangorestframework-simplejwt~=5.3.1
